requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
together>=0.1.5
pydantic>=2.4.2 